@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4",
         "composer/installers": "^1.9",
         "cweagans/composer-patches": "^1.7",
         "drupal/core-composer-scaffold": "^8.8.1",
@@ -18,6 +18,7 @@
         "drupal/media_alias_display": "^1.0",
         "drupal/media_download_path": "^2.1",
         "drupal/pathauto": "^1.8",
+        "drupal/search_api_solr": "^4.2",
         "drush-ops/behat-drush-endpoint": "^9.3",
         "drush/drush": "^8.4.5",
         "pantheon-systems/drupal-integrations": "^8",
@@ -106,7 +107,7 @@
         "preferred-install": "dist",
         "sort-packages": true,
         "platform": {
-            "php": "7.3"
+            "php": "7.4"
         }
     }
 }
