@@ -2,7 +2,7 @@
     "name":         "behat/mink-goutte-driver",
     "description":  "Goutte driver for Mink framework",
     "keywords":     ["goutte", "headless", "testing", "browser"],
-    "homepage":     "http://mink.behat.org/",
+    "homepage":     "https://mink.behat.org/",
     "type":         "mink-driver",
     "license":      "MIT",
 
@@ -15,14 +15,13 @@
     ],
 
     "require": {
-        "php":                           ">=5.3.1",
-        "behat/mink":                    "~1.6@dev",
+        "php":                           ">=5.4",
         "behat/mink-browserkit-driver":  "~1.2@dev",
         "fabpot/goutte":                 "~1.0.4|~2.0|~3.1"
     },
 
     "require-dev": {
-        "symfony/phpunit-bridge": "~2.7|~3.0"
+        "mink/driver-testsuite": "dev-master"
     },
 
     "autoload": {
@@ -39,7 +38,7 @@
 
     "extra": {
         "branch-alias": {
-            "dev-master": "1.2.x-dev"
+            "dev-master": "1.x-dev"
         }
     }
 }
