@@ -8,29 +8,41 @@
     "authors": [
         {
             "name": "See GitHub contributors",
-            "homepage": "https://github.com/basdenooijer/solarium/contributors"
+            "homepage": "https://github.com/solariumphp/solarium/contributors"
         }
     ],
     "require": {
-        "php": ">=5.3.2",
-        "symfony/event-dispatcher": "~2.3|~3.0"
+        "php": "^7.3 || ^8.0",
+        "ext-json": "*",
+        "psr/event-dispatcher": "^1.0",
+        "psr/http-client": "^1.0",
+        "psr/http-factory": "^1.0",
+        "symfony/event-dispatcher-contracts": "^1.0 || ^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~3.7",
-        "squizlabs/php_codesniffer": "~1.4",
-        "zendframework/zendframework1": "~1.12",
-        "satooshi/php-coveralls": "~1.0",
-        "guzzlehttp/guzzle": "^3.8 || ^6.2"
+        "escapestudios/symfony2-coding-standard": "^3.11",
+        "guzzlehttp/guzzle": "^7.2",
+        "nyholm/psr7": "^1.2",
+        "php-http/guzzle7-adapter": "^0.1",
+        "phpstan/extension-installer": "^1.0",
+        "phpstan/phpstan": "^0.12",
+        "phpstan/phpstan-phpunit": "^0.12",
+        "phpunit/phpunit": "^9.5",
+        "roave/security-advisories": "dev-master",
+        "symfony/event-dispatcher": "^4.3 || ^5.0"
     },
-    "suggest": {
-        "minimalcode/search": "Query builder compatible with Solarium, allows simplified solr-query handling"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-develop": "3.3.x-dev"
-        }
+    "prefer-stable": true,
+    "config": {
+        "sort-packages": true
     },
     "autoload": {
-        "psr-0": { "Solarium\\": "library/" }
+        "psr-4": {
+            "Solarium\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Solarium\\Tests\\": "tests/"
+        }
     }
 }
