@@ -2,26 +2,31 @@
 
 namespace Drupal\search_api_solr\SolrConnector;
 
+use Drupal\Component\EventDispatcher\ContainerAwareEventDispatcher;
 use Drupal\Component\Serialization\Json;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Link;
 use Drupal\Core\Plugin\PluginFormInterface;
 use Drupal\Core\Url;
+use Drupal\search_api\LoggerTrait;
 use Drupal\search_api\Plugin\ConfigurablePluginBase;
 use Drupal\search_api\Plugin\PluginFormTrait;
 use Drupal\search_api_solr\SearchApiSolrException;
+use Drupal\search_api_solr\Solarium\Autocomplete\Query as AutocompleteQuery;
 use Drupal\search_api_solr\SolrConnectorInterface;
 use Solarium\Client;
+use Solarium\Core\Client\Adapter\Curl;
+use Solarium\Core\Client\Adapter\Http;
+use Solarium\Core\Client\Adapter\TimeoutAwareInterface;
 use Solarium\Core\Client\Endpoint;
 use Solarium\Core\Client\Request;
 use Solarium\Core\Client\Response;
-use Solarium\Core\Query\Helper;
 use Solarium\Core\Query\QueryInterface;
 use Solarium\Exception\HttpException;
 use Solarium\QueryType\Extract\Result as ExtractResult;
-use Solarium\QueryType\Update\Query\Query as UpdateQuery;
 use Solarium\QueryType\Select\Query\Query;
-use Symfony\Component\DependencyInjection\ContainerInterface;
+use Solarium\QueryType\Update\Query\Query as UpdateQuery;
+use ZipStream\ZipStream;
 
 /**
  * Defines a base class for Solr connector plugins.
@@ -56,10 +61,12 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
     submitConfigurationForm as traitSubmitConfigurationForm;
   }
 
+  use LoggerTrait;
+
   /**
    * The event dispatcher.
    *
-   * @var \Symfony\Component\EventDispatcher\EventDispatcherInterface
+   * @var \Drupal\Component\EventDispatcher\ContainerAwareEventDispatcher
    */
   protected $eventDispatcher;
 
@@ -73,12 +80,9 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * {@inheritdoc}
    */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    $plugin = parent::create($container, $configuration, $plugin_id, $plugin_definition);
-
-    $plugin->eventDispatcher = $container->get('event_dispatcher');
-
-    return $plugin;
+  public function setEventDispatcher(ContainerAwareEventDispatcher $eventDispatcher) : SolrConnectorInterface {
+    $this->eventDispatcher = $eventDispatcher;
+    return $this;
   }
 
   /**
@@ -88,64 +92,87 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
     return [
       'scheme' => 'http',
       'host' => 'localhost',
-      'port' => '8983',
-      'path' => '/solr',
+      'port' => 8983,
+      'path' => '/',
       'core' => '',
       'timeout' => 5,
-      'index_timeout' => 5,
-      'optimize_timeout' => 10,
+      self::INDEX_TIMEOUT => 5,
+      self::OPTIMIZE_TIMEOUT => 10,
+      self::FINALIZE_TIMEOUT => 30,
       'solr_version' => '',
       'http_method' => 'AUTO',
       'commit_within' => 1000,
+      'jmx' => FALSE,
+      'jts' => FALSE,
+      'solr_install_dir' => '',
+      'skip_schema_check' => FALSE,
     ];
   }
 
   /**
    * {@inheritdoc}
    */
+  public function setConfiguration(array $configuration) {
+    $configuration['port'] = (int) $configuration['port'];
+    $configuration['timeout'] = (int) $configuration['timeout'];
+    $configuration[self::INDEX_TIMEOUT] = (int) $configuration[self::INDEX_TIMEOUT];
+    $configuration[self::OPTIMIZE_TIMEOUT] = (int) $configuration[self::OPTIMIZE_TIMEOUT];
+    $configuration[self::FINALIZE_TIMEOUT] = (int) $configuration[self::FINALIZE_TIMEOUT];
+    $configuration['commit_within'] = (int) $configuration['commit_within'];
+    $configuration['jmx'] = (bool) $configuration['jmx'];
+    $configuration['jts'] = (bool) $configuration['jts'];
+    $configuration['skip_schema_check'] = (bool) $configuration['skip_schema_check'];
+
+    parent::setConfiguration($configuration);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
   public function buildConfigurationForm(array $form, FormStateInterface $form_state) {
-    $form['scheme'] = array(
+    $form['scheme'] = [
       '#type' => 'select',
       '#title' => $this->t('HTTP protocol'),
       '#description' => $this->t('The HTTP protocol to use for sending queries.'),
       '#default_value' => isset($this->configuration['scheme']) ? $this->configuration['scheme'] : 'http',
-      '#options' => array(
+      '#options' => [
         'http' => 'http',
         'https' => 'https',
-      ),
-    );
+      ],
+    ];
 
-    $form['host'] = array(
+    $form['host'] = [
       '#type' => 'textfield',
       '#title' => $this->t('Solr host'),
       '#description' => $this->t('The host name or IP of your Solr server, e.g. <code>localhost</code> or <code>www.example.com</code>.'),
       '#default_value' => isset($this->configuration['host']) ? $this->configuration['host'] : '',
       '#required' => TRUE,
-    );
+    ];
 
-    $form['port'] = array(
+    $form['port'] = [
       '#type' => 'textfield',
       '#title' => $this->t('Solr port'),
       '#description' => $this->t('The Jetty example server is at port 8983, while Tomcat uses 8080 by default.'),
       '#default_value' => isset($this->configuration['port']) ? $this->configuration['port'] : '',
       '#required' => TRUE,
-    );
+    ];
 
-    $form['path'] = array(
+    $form['path'] = [
       '#type' => 'textfield',
       '#title' => $this->t('Solr path'),
       '#description' => $this->t('The path that identifies the Solr instance to use on the server.'),
-      '#default_value' => isset($this->configuration['path']) ? $this->configuration['path'] : '',
-    );
+      '#default_value' => isset($this->configuration['path']) ? $this->configuration['path'] : '/',
+    ];
 
-    $form['core'] = array(
+    $form['core'] = [
       '#type' => 'textfield',
       '#title' => $this->t('Solr core'),
       '#description' => $this->t('The name that identifies the Solr core to use on the server.'),
       '#default_value' => isset($this->configuration['core']) ? $this->configuration['core'] : '',
-    );
+      '#required' => TRUE,
+    ];
 
-    $form['timeout'] = array(
+    $form['timeout'] = [
       '#type' => 'number',
       '#min' => 1,
       '#max' => 180,
@@ -153,68 +180,109 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
       '#description' => $this->t('The timeout in seconds for search queries sent to the Solr server.'),
       '#default_value' => isset($this->configuration['timeout']) ? $this->configuration['timeout'] : 5,
       '#required' => TRUE,
-    );
+    ];
 
-    $form['index_timeout'] = array(
+    $form[self::INDEX_TIMEOUT] = [
       '#type' => 'number',
       '#min' => 1,
       '#max' => 180,
       '#title' => $this->t('Index timeout'),
       '#description' => $this->t('The timeout in seconds for indexing requests to the Solr server.'),
-      '#default_value' => isset($this->configuration['index_timeout']) ? $this->configuration['index_timeout'] : 5,
+      '#default_value' => isset($this->configuration[self::INDEX_TIMEOUT]) ? $this->configuration[self::INDEX_TIMEOUT] : 5,
       '#required' => TRUE,
-    );
+    ];
 
-    $form['optimize_timeout'] = array(
+    $form[self::OPTIMIZE_TIMEOUT] = [
       '#type' => 'number',
       '#min' => 1,
       '#max' => 180,
       '#title' => $this->t('Optimize timeout'),
       '#description' => $this->t('The timeout in seconds for background index optimization queries on a Solr server.'),
-      '#default_value' => isset($this->configuration['optimize_timeout']) ? $this->configuration['optimize_timeout'] : 10,
+      '#default_value' => isset($this->configuration[self::OPTIMIZE_TIMEOUT]) ? $this->configuration[self::OPTIMIZE_TIMEOUT] : 10,
       '#required' => TRUE,
-    );
+    ];
 
-    $form['commit_within'] = array(
+    $form[self::FINALIZE_TIMEOUT] = [
+      '#type' => 'number',
+      '#min' => 1,
+      '#max' => 180,
+      '#title' => $this->t('Finalize timeout'),
+      '#description' => $this->t('The timeout in seconds for index finalization queries on a Solr server.'),
+      '#default_value' => isset($this->configuration[self::FINALIZE_TIMEOUT]) ? $this->configuration[self::FINALIZE_TIMEOUT] : 30,
+      '#required' => TRUE,
+    ];
+
+    $form['commit_within'] = [
       '#type' => 'number',
       '#min' => 0,
       '#title' => $this->t('Commit within'),
       '#description' => $this->t('The limit in milliseconds within a (soft) commit on Solr is forced after any updating the index in any way. Setting the value to "0" turns off this dynamic enforcement and lets Solr behave like configured solrconf.xml.'),
       '#default_value' => isset($this->configuration['commit_within']) ? $this->configuration['commit_within'] : 1000,
       '#required' => TRUE,
-    );
+    ];
 
-    $form['workarounds'] = array(
-      '#type' => 'fieldset',
+    $form['workarounds'] = [
+      '#type' => 'details',
       '#title' => $this->t('Connector Workarounds'),
-      '#collapsible' => TRUE,
-      '#collapsed' => TRUE,
-    );
+    ];
 
-    $form['workarounds']['solr_version'] = array(
+    $form['workarounds']['solr_version'] = [
       '#type' => 'select',
       '#title' => $this->t('Solr version override'),
       '#description' => $this->t('Specify the Solr version manually in case it cannot be retrived automatically. The version can be found in the Solr admin interface under "Solr Specification Version" or "solr-spec"'),
-      '#options' => array(
+      '#options' => [
         '' => $this->t('Determine automatically'),
-        '4' => '4.x',
-        '5' => '5.x',
         '6' => '6.x',
-      ),
-      '#default_value' => isset($this->configuration['solr_version']) ? $this->configuration['solr_version'] : '',
-    );
+        '7' => '7.x',
+        '8' => '8.x',
+      ],
+      '#default_value' => $this->configuration['solr_version'] ?? '',
+    ];
 
-    $form['workarounds']['http_method'] = array(
+    $form['workarounds']['http_method'] = [
       '#type' => 'select',
       '#title' => $this->t('HTTP method'),
       '#description' => $this->t('The HTTP method to use for sending queries. GET will often fail with larger queries, while POST should not be cached. AUTO will use GET when possible, and POST for queries that are too large.'),
-      '#default_value' => isset($this->configuration['http_method']) ? $this->configuration['http_method'] : 'AUTO',
-      '#options' => array(
+      '#default_value' => $this->configuration['http_method'] ?? 'AUTO',
+      '#options' => [
         'AUTO' => $this->t('AUTO'),
         'POST' => 'POST',
         'GET' => 'GET',
-      ),
-    );
+      ],
+    ];
+
+    $form['workarounds']['skip_schema_check'] = [
+      '#type' => 'checkbox',
+      '#title' => $this->t('Skip schema verification'),
+      '#description' => $this->t('Skip the automatic check for schema-compatibillity. Use this override if you are seeing an error-message about an incompatible schema.xml configuration file, and you are sure the configuration is compatible.'),
+      '#default_value' => $this->configuration['skip_schema_check'] ?? FALSE,
+    ];
+
+    $form['advanced'] = [
+      '#type' => 'details',
+      '#title' => $this->t('Advanced server configuration'),
+    ];
+
+    $form['advanced']['jmx'] = [
+      '#type' => 'checkbox',
+      '#title' => $this->t('Enable JMX'),
+      '#description' => $this->t('Enable JMX based monitoring.'),
+      '#default_value' => $this->configuration['jmx'] ?? FALSE,
+    ];
+
+    $form['advanced']['jts'] = [
+      '#type' => 'checkbox',
+      '#title' => $this->t('Enable JTS'),
+      '#description' => $this->t('Enable JTS (java topographic suite). Be sure to follow instructions in last solr reference guide about how to use spatial search.'),
+      '#default_value' => $this->configuration['jts'] ?? FALSE,
+    ];
+
+    $form['advanced']['solr_install_dir'] = [
+      '#type' => 'textfield',
+      '#title' => $this->t('solr.install.dir'),
+      '#description' => $this->t('The path where Solr is installed on the server, relative to the configuration or absolute. Some examples are "../../.." for Solr downloaded from apache.org, "/usr/local/opt/solr" for installations via homebrew on macOS or "/opt/solr" for some linux distributions and for the official Solr docker container. If you use different systems for development, testing and production you can use drupal config overwrites to adjust the value per environment or adjust the generated solrcore.properties per environment or use java virtual machine options (-D) to set the property. Modern Solr installations should set that virtual machine option correctly in their start script by themselves. In this case this field should be left empty!'),
+      '#default_value' => isset($this->configuration['solr_install_dir']) ? $this->configuration['solr_install_dir'] : '',
+    ];
 
     return $form;
   }
@@ -231,13 +299,14 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
       $form_state->setError($form['path'], $this->t('If provided the path has to start with "/".'));
     }
     if (!empty($values['core']) && strpos($values['core'], '/') === 0) {
-      $form_state->setError($form['core'], $this->t('The core must not start with "/".'));
+      $form_state->setError($form['core'], $this->t('Core or collection must not start with "/".'));
     }
 
     if (!$form_state->hasAnyErrors()) {
       // Try to orchestrate a server link from form values.
-      $solr = new Client(NULL, $this->eventDispatcher);
-      $solr->createEndpoint($values + ['key' => 'core'], TRUE);
+      $values_copied = $values;
+      $solr = $this->createClient($values_copied);
+      $solr->createEndpoint($values_copied + ['key' => 'search_api_solr'], TRUE);
       try {
         $this->getServerLink();
       }
@@ -261,9 +330,13 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
     foreach ($values['workarounds'] as $key => $value) {
       $form_state->setValue($key, $value);
     }
+    foreach ($values['advanced'] as $key => $value) {
+      $form_state->setValue($key, $value);
+    }
 
     // Clean-up the form to avoid redundant entries in the stored configuration.
     $form_state->unsetValue('workarounds');
+    $form_state->unsetValue('advanced');
 
     $this->traitSubmitConfigurationForm($form, $form_state);
   }
@@ -273,27 +346,32 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
    */
   protected function connect() {
     if (!$this->solr) {
-      $this->solr = new Client(NULL, $this->eventDispatcher);
-      $this->solr->createEndpoint($this->configuration + ['key' => 'core'], TRUE);
-      $this->attachServerEndpoint();
+      $configuration = $this->configuration;
+      $this->solr = $this->createClient($configuration);
+      $this->solr->createEndpoint($configuration + ['key' => 'search_api_solr'], TRUE);
     }
   }
 
   /**
-   * Attaches an endpoint to the Solr connection to communicate with the server.
-   *
-   * This endpoint is different from the core endpoint which is the default one.
-   * The default endpoint for the core is used to communicate with the index.
-   * But for some administrative tasks the server itself needs to be contacted.
-   * This function is meant to be overwritten as soon as we deal with Solr
-   * service provider specific implementations of SolrHelper.
+   * Create a Client.
    */
-  protected function attachServerEndpoint() {
-    $this->connect();
-    $configuration = $this->configuration;
-    $configuration['core'] = NULL;
-    $configuration['key'] = 'server';
-    $this->solr->createEndpoint($configuration);
+  protected function createClient(array &$configuration) {
+    // @todo For backward compatibility we didn't rename 'timeout' yet. We
+    // should do so in an update hook.
+    $configuration[self::QUERY_TIMEOUT] = $configuration['timeout'] ?? 5;
+    unset($configuration['timeout']);
+
+    $adapter = extension_loaded('curl') ? new Curl() : new Http();
+    $adapter->setTimeout($configuration[self::QUERY_TIMEOUT]);
+
+    return new Client($adapter, $this->eventDispatcher);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function isCloud() {
+    return FALSE;
   }
 
   /**
@@ -301,8 +379,8 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
    */
   protected function getServerUri() {
     $this->connect();
-    $url_path = $this->solr->getEndpoint('server')->getBaseUri();
-    if ($this->configuration['host'] == 'localhost' && !empty($_SERVER['SERVER_NAME'])) {
+    $url_path = $this->solr->getEndpoint()->getServerUri();
+    if ($this->configuration['host'] === 'localhost' && !empty($_SERVER['SERVER_NAME'])) {
       $url_path = str_replace('localhost', $_SERVER['SERVER_NAME'], $url_path);
     }
 
@@ -323,7 +401,7 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
    * {@inheritdoc}
    */
   public function getCoreLink() {
-    $url_path = $this->getServerUri() . '#/' . $this->configuration['core'];
+    $url_path = $this->getServerUri() . 'solr/#/' . $this->configuration['core'];
     $url = Url::fromUri($url_path);
 
     return Link::fromTextAndUrl($url_path, $url);
@@ -340,14 +418,32 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
       // corresponding full version string.
       $min_version = ['0', '0', '0'];
       $version = implode('.', explode('.', $this->configuration['solr_version']) + $min_version);
-      return '4.0.0' === $version ? '4.5.0' : $version;
+      switch ($version) {
+        case '3.0.0':
+          // 3.6.0 is the minimum supported Solr 3 version by the
+          // search_api_solr_legacy module.
+          $version = '3.6.0';
+          break;
+        case '4.0.0':
+          // 4.5.0 is the minimum supported Solr 4 version by the
+          // search_api_solr_legacy module.
+          $version = '4.5.0';
+          break;
+        case '6.0.0':
+          // 6.4.0 is the minimum supported Solr version. Earlier Solr 6
+          // versions should run in Solr 5 compatibility mode using the
+          // search_api_solr_legacy module.
+          $version = '6.4.0';
+          break;
+      }
+      return $version;
     }
 
     $info = [];
     try {
       $info = $this->getCoreInfo();
     }
-    catch (SearchApiSolrException $e) {
+    catch (\Exception $e) {
       try {
         $info = $this->getServerInfo();
       }
@@ -357,7 +453,11 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
 
     // Get our solr version number.
     if (isset($info['lucene']['solr-spec-version'])) {
-      return $info['lucene']['solr-spec-version'];
+      // Some Solr distributions or docker images append additional info to the
+      // version number, for example the build date: 3.6.2.2012.12.18.19.52.27.
+      if (preg_match('/^(\d+\.\d+\.\d+)/', $info['lucene']['solr-spec-version'], $matches)) {
+        return $matches[1];
+      }
     }
 
     return '0.0.0';
@@ -366,9 +466,9 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * {@inheritdoc}
    */
-  public function getSolrMajorVersion($version = '') {
-    list($major, ,) = explode('.', $version ?: $this->getSolrVersion());
-    return $major;
+  public function getSolrMajorVersion($version = ''): int {
+    [$major] = explode('.', $version ?: $this->getSolrVersion());
+    return (int) $major;
   }
 
   /**
@@ -382,7 +482,7 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
    * {@inheritdoc}
    */
   public function getLuceneMatchVersion($version = '') {
-    list($major, $minor,) = explode('.', $version ?: $this->getSolrVersion());
+    [$major, $minor] = explode('.', $version ?: $this->getSolrVersion());
     return $major . '.' . $minor;
   }
 
@@ -390,21 +490,31 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
    * {@inheritdoc}
    */
   public function getServerInfo($reset = FALSE) {
-    return $this->getDataFromHandler('server', 'admin/info/system', $reset);
+    $this->useTimeout();
+    return $this->getDataFromHandler('admin/info/system', $reset);
   }
 
   /**
    * {@inheritdoc}
    */
   public function getCoreInfo($reset = FALSE) {
-    return $this->getDataFromHandler('core', 'admin/system', $reset);
+    $this->useTimeout();
+    return $this->getDataFromHandler($this->configuration['core'] . '/admin/system', $reset);
   }
 
   /**
    * {@inheritdoc}
    */
   public function getLuke() {
-    return $this->getDataFromHandler('core', 'admin/luke', TRUE);
+    $this->useTimeout();
+    return $this->getDataFromHandler($this->configuration['core'] . '/admin/luke', TRUE);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getConfigSetName(): ?string {
+    return NULL;
   }
 
   /**
@@ -423,75 +533,90 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   }
 
   /**
-   * Gets data from a Solr endpoint using a given handler.
-   *
-   * @param $endpoint_name
-   * @param $handler
-   * @param bool $reset
-   *   If TRUE the server will be asked regardless if a previous call is cached.
-   *
-   * @return object
-   *   A response object with system information.
+   * {@inheritdoc}
    */
-  protected function getDataFromHandler($endpoint_name, $handler, $reset = FALSE) {
-    static $previous_calls = [];
-
-    $this->connect();
-
-    $endpoint = $this->solr->getEndpoint($endpoint_name);
-    $endpoint_uri = $endpoint->getBaseUri();
-    $state_key = 'search_api_solr.endpoint.data';
-    $state = \Drupal::state();
-    $endpoint_data = $state->get($state_key);
-
-    if (!isset($previous_calls[$endpoint_uri][$handler]) || $reset) {
-      // Don't retry multiple times in case of an exception.
-      $previous_calls[$endpoint_name] = TRUE;
-
-      if (!is_array($endpoint_data) || !isset($endpoint_data[$endpoint_uri][$handler]) || $reset) {
-        // @todo Finish https://github.com/solariumphp/solarium/pull/155 and stop
-        // abusing the ping query for this.
-        $query = $this->solr->createPing(array('handler' => $handler));
-        $endpoint_data[$endpoint_uri][$handler] = $this->execute($query, $endpoint)->getData();
-        $state->set($state_key, $endpoint_data);
-      }
-    }
-
-    return $endpoint_data[$endpoint_uri][$handler];
+  public function getSchemaTargetedSolrBranch($reset = FALSE) {
+    $parts = explode('-', $this->getSchemaVersionString($reset));
+    return $parts[3];
   }
 
   /**
    * {@inheritdoc}
    */
-  public function pingCore() {
-    return $this->doPing();
+  public function isJumpStartConfigSet(bool $reset = FALSE): bool {
+    $parts = explode('-', $this->getSchemaVersionString($reset));
+    return (bool) ($parts[4] ?? 0);
+  }
+
+  /**
+   * Gets data from a Solr endpoint using a given handler.
+   *
+   * @param string $handler
+   *   The handler used for the API query.
+   * @param bool $reset
+   *   If TRUE the server will be asked regardless if a previous call is cached.
+   *
+   * @return array
+   *   Response data with system information.
+   *
+   * @throws \Drupal\search_api_solr\SearchApiSolrException
+   */
+  protected function getDataFromHandler($handler, $reset = FALSE) {
+    static $previous_calls = [];
+
+    $this->connect();
+
+    // We keep the results in a state instead of a cache because we want to
+    // access parts of this data even if Solr is temporarily not reachable and
+    // caches have been cleared.
+    $state_key = 'search_api_solr.endpoint.data';
+    $state = \Drupal::state();
+    $endpoint_data = $state->get($state_key);
+    $server_uri = $this->getServerUri();
+
+    if (!isset($previous_calls[$server_uri][$handler]) || !isset($endpoint_data[$server_uri][$handler]) || $reset) {
+      // Don't retry multiple times in case of an exception.
+      $previous_calls[$server_uri][$handler] = TRUE;
+
+      if (!is_array($endpoint_data) || !isset($endpoint_data[$server_uri][$handler]) || $reset) {
+        $query = $this->solr->createApi([
+          'handler' => $handler,
+          'version' => Request::API_V1,
+        ]);
+        $endpoint_data[$server_uri][$handler] = $this->execute($query)->getData();
+        $state->set($state_key, $endpoint_data);
+      }
+    }
+
+    return $endpoint_data[$server_uri][$handler];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function pingCore(array $options = []) {
+    return $this->pingEndpoint(NULL, $options);
   }
 
   /**
    * {@inheritdoc}
    */
   public function pingServer() {
-    return $this->doPing(['handler' => 'admin/info/system'], 'server');
+    return $this->getServerInfo(TRUE);
   }
 
   /**
-   * Pings the Solr server to tell whether it can be accessed.
-   *
-   * @param string $endpoint_name
-   *   The endpoint to be pinged on the Solr server.
-   *
-   * @return mixed
-   *   The latency in milliseconds if the core can be accessed,
-   *   otherwise FALSE.
+   * {@inheritdoc}
    */
-  protected function doPing($options = [], $endpoint_name = 'core') {
+  public function pingEndpoint(?Endpoint $endpoint = NULL, array $options = []) {
     $this->connect();
-    // Default is ['handler' => 'admin/ping'].
+    $this->useTimeout(self::QUERY_TIMEOUT, $endpoint);
+
     $query = $this->solr->createPing($options);
 
     try {
       $start = microtime(TRUE);
-      $result = $this->solr->execute($query, $endpoint_name);
+      $result = $this->solr->execute($query, $endpoint);
       if ($result->getResponse()->getStatusCode() == 200) {
         // Add 1 µs to the ping time so we never return 0.
         return (microtime(TRUE) - $start) + 1E-6;
@@ -509,8 +634,9 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
    */
   public function getStatsSummary() {
     $this->connect();
+    $this->useTimeout();
 
-    $summary = array(
+    $summary = [
       '@pending_docs' => '',
       '@autocommit_time_seconds' => '',
       '@autocommit_time' => '',
@@ -520,31 +646,46 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
       '@schema_version' => '',
       '@core_name' => '',
       '@index_size' => '',
-    );
+    ];
 
     $query = $this->solr->createPing();
     $query->setResponseWriter(Query::WT_PHPS);
     $query->setHandler('admin/mbeans?stats=true');
     $stats = $this->execute($query)->getData();
     if (!empty($stats)) {
-      $update_handler_stats = $stats['solr-mbeans']['UPDATEHANDLER']['updateHandler']['stats'];
-      $summary['@pending_docs'] = (int) $update_handler_stats['docsPending'];
-      $max_time = (int) $update_handler_stats['autocommit maxTime'];
-      // Convert to seconds.
-      $summary['@autocommit_time_seconds'] = $max_time / 1000;
-      $summary['@autocommit_time'] = \Drupal::service('date.formatter')->formatInterval($max_time / 1000);
-      $summary['@deletes_by_id'] = (int) $update_handler_stats['deletesById'];
-      $summary['@deletes_by_query'] = (int) $update_handler_stats['deletesByQuery'];
-      $summary['@deletes_total'] = $summary['@deletes_by_id'] + $summary['@deletes_by_query'];
-      $summary['@schema_version'] = $this->getSchemaVersionString(TRUE);
-      $summary['@core_name'] = $stats['solr-mbeans']['CORE']['core']['stats']['coreName'];
-      if (version_compare($this->getSolrVersion(TRUE), '6.4', '>=')) {
-        // @see https://issues.apache.org/jira/browse/SOLR-3990
-        $summary['@index_size'] = $stats['solr-mbeans']['CORE']['core']['stats']['size'];
+      $solr_version = $this->getSolrVersion(TRUE);
+      $max_time = -1;
+      if (version_compare($solr_version, '7.0', '>=')) {
+        $update_handler_stats = $stats['solr-mbeans']['UPDATE']['updateHandler']['stats'];
+        $summary['@pending_docs'] = (int) $update_handler_stats['UPDATE.updateHandler.docsPending'];
+        if (isset($update_handler_stats['UPDATE.updateHandler.softAutoCommitMaxTime'])) {
+          $max_time = (int) $update_handler_stats['UPDATE.updateHandler.softAutoCommitMaxTime'];
+        }
+        $summary['@deletes_by_id'] = (int) $update_handler_stats['UPDATE.updateHandler.deletesById'];
+        $summary['@deletes_by_query'] = (int) $update_handler_stats['UPDATE.updateHandler.deletesByQuery'];
+        $summary['@core_name'] = $stats['solr-mbeans']['CORE']['core']['stats']['CORE.coreName'] ?? $this->t('No information available.');
+        $summary['@index_size'] = $stats['solr-mbeans']['CORE']['core']['stats']['INDEX.size'] ?? $this->t('No information available.');
       }
       else {
-        $summary['@index_size'] = $stats['solr-mbeans']['QUERYHANDLER']['/replication']['stats']['indexSize'];
+        $update_handler_stats = $stats['solr-mbeans']['UPDATEHANDLER']['updateHandler']['stats'];
+        $summary['@pending_docs'] = (int) $update_handler_stats['docsPending'];
+        $max_time = (int) $update_handler_stats['autocommit maxTime'];
+        $summary['@deletes_by_id'] = (int) $update_handler_stats['deletesById'];
+        $summary['@deletes_by_query'] = (int) $update_handler_stats['deletesByQuery'];
+        $summary['@core_name'] = $stats['solr-mbeans']['CORE']['core']['stats']['coreName'] ?? $this->t('No information available.');;
+        if (version_compare($solr_version, '6.4', '>=')) {
+          // @see https://issues.apache.org/jira/browse/SOLR-3990
+          $summary['@index_size'] = $stats['solr-mbeans']['CORE']['core']['stats']['size'] ?? $this->t('No information available.');;
+        }
+        else {
+          $summary['@index_size'] = $stats['solr-mbeans']['QUERYHANDLER']['/replication']['stats']['indexSize'] ?? $this->t('No information available.');;
+        }
       }
+
+      $summary['@autocommit_time_seconds'] = $max_time / 1000;
+      $summary['@autocommit_time'] = \Drupal::service('date.formatter')->formatInterval($max_time / 1000);
+      $summary['@deletes_total'] = $summary['@deletes_by_id'] + $summary['@deletes_by_query'];
+      $summary['@schema_version'] = $this->getSchemaVersionString(TRUE);
     }
     return $summary;
   }
@@ -552,66 +693,63 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * {@inheritdoc}
    */
-  public function coreRestGet($path) {
-    return $this->restRequest('core', $path);
+  public function coreRestGet($path, ?Endpoint $endpoint = NULL) {
+    $this->useTimeout();
+    return $this->restRequest($this->configuration['core'] . '/' . ltrim($path, '/'), Request::METHOD_GET, '', $endpoint);
   }
 
   /**
    * {@inheritdoc}
    */
-  public function coreRestPost($path, $command_json = '') {
-    return $this->restRequest('core', $path, Request::METHOD_POST, $command_json);
+  public function coreRestPost($path, $command_json = '', ?Endpoint $endpoint = NULL) {
+    $this->useTimeout(self::INDEX_TIMEOUT);
+    return $this->restRequest($this->configuration['core'] . '/' . ltrim($path, '/'), Request::METHOD_POST, $command_json, $endpoint);
   }
 
   /**
    * {@inheritdoc}
    */
   public function serverRestGet($path) {
-    return $this->restRequest('server', $path);
+    $this->useTimeout();
+    return $this->restRequest($path);
   }
 
   /**
    * {@inheritdoc}
    */
   public function serverRestPost($path, $command_json = '') {
-    return $this->restRequest('server', $path, Request::METHOD_POST, $command_json);
+    $this->useTimeout(self::INDEX_TIMEOUT);
+    return $this->restRequest($path, Request::METHOD_POST, $command_json);
   }
 
   /**
    * Sends a REST request to the Solr server endpoint and returns the result.
    *
-   * @param string $endpoint_key
-   *   The endpoint that refelcts the base URI.
-   * @param string $path
-   *   The path to append to the base URI.
+   * @param string $handler
+   *   The handler used for the API query.
    * @param string $method
    *   The HTTP request method.
    * @param string $command_json
    *   The command to send encoded as JSON.
+   * @param \Solarium\Core\Client\Endpoint|null $endpoint
    *
-   * @return string
+   * @return array
    *   The decoded response.
+   *
+   * @throws \Drupal\search_api_solr\SearchApiSolrException
    */
-  protected function restRequest($endpoint_key, $path, $method = Request::METHOD_GET, $command_json = '') {
+  protected function restRequest($handler, $method = Request::METHOD_GET, $command_json = '', ?Endpoint $endpoint = NULL) {
     $this->connect();
+    $query = $this->solr->createApi([
+      'handler' => $handler,
+      'accept' => 'application/json',
+      'contenttype' => 'application/json',
+      'method' => $method,
+      'rawdata' => (Request::METHOD_POST == $method ? $command_json : NULL),
+    ]);
 
-    $request = new Request();
-    $request->setMethod($method);
-    $request->addHeader('Accept: application/json');
-    if (Request::METHOD_POST == $method) {
-      $request->addHeader('Content-type: application/json');
-      $request->setRawData($command_json);
-    }
-    $request->setHandler($path);
-
-    $endpoint = $this->solr->getEndpoint($endpoint_key);
-    $timeout = $endpoint->getTimeout();
-    // @todo Destinguish between different flavors of REST requests and use
-    //   different timeout settings.
-    $endpoint->setTimeout($this->configuration['optimize_timeout']);
-    $response = $this->executeRequest($request, $endpoint);
-    $endpoint->setTimeout($timeout);
-    $output = Json::decode($response->getBody());
+    $response = $this->execute($query, $endpoint);
+    $output = $response->getData();
     // \Drupal::logger('search_api_solr')->info(print_r($output, true));.
     if (!empty($output['errors'])) {
       throw new SearchApiSolrException('Error trying to send a REST request.' .
@@ -655,12 +793,37 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * {@inheritdoc}
    */
+  public function getSpellcheckQuery() {
+    $this->connect();
+    return $this->solr->createSpellcheck();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getSuggesterQuery() {
+    $this->connect();
+    return $this->solr->createSuggester();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getAutocompleteQuery() {
+    $this->connect();
+    $this->solr->registerQueryType('autocomplete', AutocompleteQuery::class);
+    return $this->solr->createQuery('autocomplete');
+  }
+
+  /**
+   * {@inheritdoc}
+   */
   public function getQueryHelper(QueryInterface $query = NULL) {
     if ($query) {
       return $query->getHelper();
     }
 
-    return new Helper();
+    return \Drupal::service('solarium.query_helper');
   }
 
   /**
@@ -672,7 +835,10 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   }
 
   /**
-   * @return \Solarium\Plugin\CustomizeRequest\CustomizeRequest
+   * Creates a CustomizeRequest object.
+   *
+   * @return \Solarium\Plugin\CustomizeRequest\CustomizeRequest|\Solarium\Core\Plugin\PluginInterface
+   *   The Solarium CustomizeRequest object.
    */
   protected function customizeRequest() {
     $this->connect();
@@ -682,17 +848,19 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * {@inheritdoc}
    */
-  public function search(Query $query, Endpoint $endpoint = NULL) {
+  public function search(Query $query, ?Endpoint $endpoint = NULL) {
     $this->connect();
 
     if (!$endpoint) {
-      $endpoint = $this->solr->getEndpoint('core');
+      $endpoint = $this->solr->getEndpoint();
     }
+
+    $this->useTimeout(self::QUERY_TIMEOUT, $endpoint);
 
     // Use the 'postbigrequest' plugin if no specific http method is
     // configured. The plugin needs to be loaded before the request is
     // created.
-    if ($this->configuration['http_method'] == 'AUTO') {
+    if ($this->configuration['http_method'] === 'AUTO') {
       $this->solr->getPlugin('postbigrequest');
     }
 
@@ -701,10 +869,10 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
     $request = $this->solr->createRequest($query);
 
     // Set the configured HTTP method.
-    if ($this->configuration['http_method'] == 'POST') {
+    if ($this->configuration['http_method'] === 'POST') {
       $request->setMethod(Request::METHOD_POST);
     }
-    elseif ($this->configuration['http_method'] == 'GET') {
+    elseif ($this->configuration['http_method'] === 'GET') {
       $request->setMethod(Request::METHOD_GET);
     }
 
@@ -714,24 +882,24 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * {@inheritdoc}
    */
-  public function createSearchResult(Query $query, Response $response) {
+  public function createSearchResult(QueryInterface $query, Response $response) {
     return $this->solr->createResult($query, $response);
   }
 
   /**
    * {@inheritdoc}
    */
-  public function update(UpdateQuery $query, Endpoint $endpoint = NULL) {
+  public function update(UpdateQuery $query, ?Endpoint $endpoint = NULL) {
     $this->connect();
 
     if (!$endpoint) {
-      $endpoint = $this->solr->getEndpoint('core');
+      $endpoint = $this->solr->getEndpoint();
     }
     // The default timeout is set for search queries. The configured timeout
     // might differ and needs to be set now because solarium doesn't
     // distinguish between these types.
-    $timeout = $endpoint->getTimeout();
-    $endpoint->setTimeout($this->configuration['index_timeout']);
+    $this->useTimeout(self::INDEX_TIMEOUT, $endpoint);
+
     if ($this->configuration['commit_within']) {
       // Do a commitWithin since that is automatically a softCommit since Solr 4
       // and a delayed hard commit with Solr 3.4+.
@@ -741,28 +909,48 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
       // @see http://wiki.apache.org/solr/NearRealtimeSearch
       /** @var \Solarium\Plugin\CustomizeRequest\CustomizeRequest $request */
       $request = $this->customizeRequest();
-      $request->createCustomization('id')
-        ->setType('param')
-        ->setName('commitWithin')
-        ->setValue($this->configuration['commit_within']);
+      if (!$request->getCustomization('commitWithin')) {
+        $request->createCustomization('commitWithin')
+          ->setType('param')
+          ->setName('commitWithin')
+          ->setValue($this->configuration['commit_within']);
+      }
     }
 
-    $result = $this->execute($query, $endpoint);
+    return $this->execute($query, $endpoint);
+  }
 
-    // Reset the timeout setting to the default value for search queries.
-    $endpoint->setTimeout($timeout);
 
-    return $result;
+  /**
+   * {@inheritdoc}
+   */
+  public function autocomplete(AutocompleteQuery $query, ?Endpoint $endpoint = NULL) {
+    $this->connect();
+
+    if (!$endpoint) {
+      $endpoint = $this->solr->getEndpoint();
+    }
+
+    $this->useTimeout(self::QUERY_TIMEOUT, $endpoint);
+
+    // Use the 'postbigrequest' plugin if no specific http method is
+    // configured. The plugin needs to be loaded before the request is
+    // created.
+    if ($this->configuration['http_method'] === 'AUTO') {
+      $this->solr->getPlugin('postbigrequest');
+    }
+
+    return $this->execute($query, $endpoint);
   }
 
   /**
    * {@inheritdoc}
    */
-  public function execute(QueryInterface $query, Endpoint $endpoint = NULL) {
+  public function execute(QueryInterface $query, ?Endpoint $endpoint = NULL) {
     $this->connect();
 
     if (!$endpoint) {
-      $endpoint = $this->solr->getEndpoint('core');
+      $endpoint = $this->solr->getEndpoint();
     }
 
     try {
@@ -776,11 +964,11 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * {@inheritdoc}
    */
-  public function executeRequest(Request $request, Endpoint $endpoint = NULL) {
+  public function executeRequest(Request $request, ?Endpoint $endpoint = NULL) {
     $this->connect();
 
     if (!$endpoint) {
-      $endpoint = $this->solr->getEndpoint('core');
+      $endpoint = $this->solr->getEndpoint();
     }
 
     try {
@@ -794,74 +982,207 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * Converts a HttpException in an easier to read SearchApiSolrException.
    *
+   * Connectors must not overwrite this function. Otherwise support requests are
+   * hard to handle in the issue queue. If you want to extend this function and
+   * add more sophisticated error handling, please contribute a patch to
+   * the search_api_solr project on drupal.org.
+   *
    * @param \Solarium\Exception\HttpException $e
+   *   The HttpException object.
    * @param \Solarium\Core\Client\Endpoint $endpoint
+   *   The Solarium endpoint.
    *
    * @throws \Drupal\search_api_solr\SearchApiSolrException
    */
-  protected function handleHttpException(HttpException $e, Endpoint $endpoint) {
-    $response_code = $e->getCode();
-    switch ($response_code) {
-      case 404:
-        $description = $this->t('not found');
+  final protected function handleHttpException(HttpException $e, Endpoint $endpoint) {
+    $body = $e->getBody();
+    $response_code = (int) $e->getCode();
+    switch ((string) $response_code) {
+      case '400': // Bad Request.
+        $description = 'bad request';
+        $response_decoded = Json::decode($body);
+        if ($response_decoded && isset($response_decoded['error'])) {
+          $body = $response_decoded['error']['msg'] ?? $body;
+        }
         break;
 
-      case 401:
-      case 403:
-        $description = $this->t('access denied');
+      case '404': // Not Found.
+        $description = 'not found';
+        break;
+
+      case '401': // Unauthorized.
+      case '403': // Forbidden.
+        $description = 'access denied';
+        break;
+
+      case '500': // Internal Server Error.
+      case '0':
+        $description = 'internal Solr server error';
         break;
 
       default:
-        $description = $this->t('unreachable');
+        $description = 'unreachable or returned unexpected response code';
     }
-    throw new SearchApiSolrException($this->t('Solr endpoint @endpoint @description.', ['@endpoint' => $endpoint->getBaseUri(), '@description' => $description]), $response_code, $e);
+    throw new SearchApiSolrException(sprintf('Solr endpoint %s %s (code: %d, body: %s, message: %s).', $this->getEndpointUri($endpoint), $description, $response_code, $body, $e->getMessage()), $response_code, $e);
+  }
+
+  /**
+   * Gets a string representation of the endpoint URI.
+   *
+   * Could be overwritten by other connectors according to their needs.
+   *
+   * @param \Solarium\Core\Client\Endpoint $endpoint
+   *
+   * @return string
+   */
+  protected function getEndpointUri(Endpoint $endpoint): string {
+    return $endpoint->getServerUri();
   }
 
   /**
    * {@inheritdoc}
    */
-  public function optimize(Endpoint $endpoint = NULL) {
+  public function optimize(?Endpoint $endpoint = NULL) {
     $this->connect();
 
     if (!$endpoint) {
-      $endpoint = $this->solr->getEndpoint('core');
+      $endpoint = $this->solr->getEndpoint();
     }
     // The default timeout is set for search queries. The configured timeout
     // might differ and needs to be set now because solarium doesn't
     // distinguish between these types.
-    $timeout = $endpoint->getTimeout();
-    $endpoint->setTimeout($this->configuration['optimize_timeout']);
+    $this->useTimeout(self::OPTIMIZE_TIMEOUT, $endpoint);
 
     $update_query = $this->solr->createUpdate();
     $update_query->addOptimize(TRUE, FALSE);
 
     $this->execute($update_query, $endpoint);
-
-    // Reset the timeout setting to the default value for search queries.
-    $endpoint->setTimeout($timeout);
   }
 
   /**
    * {@inheritdoc}
    */
-  public function extract(QueryInterface $query) {
-    return $this->execute($query);
+  public function adjustTimeout(int $seconds, string $timeout = self::QUERY_TIMEOUT, ?Endpoint &$endpoint = NULL): int {
+    $this->connect();
+
+    if (!$endpoint) {
+      $endpoint = $this->solr->getEndpoint();
+    }
+
+    $previous_timeout = $endpoint->getOption($timeout);
+    $options = $endpoint->getOptions();
+    $options[$timeout] = $seconds;
+    $endpoint = new Endpoint($options);
+    return $previous_timeout;
+  }
+
+  /**
+   * Set the timeout.
+   *
+   * @param string $timeout
+   *   (optional) The configured timeout to use. Default is self::QUERY_TIMEOUT.
+   * @param \Solarium\Core\Client\Endpoint|null $endpoint
+   *   (optional) The Solarium endpoint object.
+   * @return mixed
+   */
+  protected function useTimeout(string $timeout = self::QUERY_TIMEOUT, ?Endpoint $endpoint = NULL) {
+    $this->connect();
+
+    if (!$endpoint) {
+      $endpoint = $this->solr->getEndpoint();
+    }
+    $seconds = $endpoint->getOption($timeout);
+    if ($seconds) {
+      $adapter = $this->solr->getAdapter();
+      if ($adapter instanceof TimeoutAwareInterface) {
+        $adapter->setTimeout($seconds);
+      }
+      else {
+        $this->getLogger()->warning('The function SolrConnectorPluginBase::useTimeout() has no effect because you use a HTTP adapter that is not implementing TimeoutAwareInterface. You need to adjust your SolrConnector accordingly.');
+      }
+    }
   }
 
   /**
    * {@inheritdoc}
    */
-  public function getContentFromExtractResult(ExtractResult $result, $filepath){
-    $response = $result->getResponse();
-    $json_data = $response->getBody();
-    $array_data = Json::decode($json_data);
-    return $array_data[$filepath];
+  public function getTimeout(?Endpoint $endpoint = NULL) {
+    $this->connect();
+
+    if (!$endpoint) {
+      $endpoint = $this->solr->getEndpoint();
+    }
+
+    return $endpoint->getOption(self::QUERY_TIMEOUT);
   }
 
   /**
    * {@inheritdoc}
    */
-  public function getEndpoint($key = 'core') {
+  public function getIndexTimeout(?Endpoint $endpoint = NULL) {
+    $this->connect();
+
+    if (!$endpoint) {
+      $endpoint = $this->solr->getEndpoint();
+    }
+
+    return $endpoint->getOption(self::INDEX_TIMEOUT);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getOptimizeTimeout(?Endpoint $endpoint = NULL) {
+    $this->connect();
+
+    if (!$endpoint) {
+      $endpoint = $this->solr->getEndpoint();
+    }
+
+    return $endpoint->getOption(self::OPTIMIZE_TIMEOUT);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getFinalizeTimeout(?Endpoint $endpoint = NULL) {
+    $this->connect();
+
+    if (!$endpoint) {
+      $endpoint = $this->solr->getEndpoint();
+    }
+
+    return $endpoint->getOption(self::FINALIZE_TIMEOUT);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function extract(QueryInterface $query, ?Endpoint $endpoint = NULL) {
+    $this->useTimeout(self::INDEX_TIMEOUT, $endpoint);
+    return $this->execute($query, $endpoint);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getContentFromExtractResult(ExtractResult $result, $filepath) {
+    $array_data = $result->getData();
+
+    if (isset($array_data[basename($filepath)])) {
+      return $array_data[basename($filepath)];
+    }
+
+    // In most (or every) cases when an error happens we won't reach that point,
+    // because a Solr exception is already pased through. Anyway, this exception
+    // will be thrown if the solarium library surprises us again. ;-)
+    throw new SearchApiSolrException('Unable to find extracted files within the Solr response body.');
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getEndpoint($key = 'search_api_solr') {
     $this->connect();
     return $this->solr->getEndpoint($key);
   }
@@ -869,12 +1190,22 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
   /**
    * {@inheritdoc}
    */
+  public function createEndpoint(string $key, array $additional_configuration = []) {
+    $this->connect();
+    $configuration = ['key' => $key, self::QUERY_TIMEOUT => $this->configuration['timeout']] + $additional_configuration + $this->configuration;
+    unset($configuration['timeout']);
+
+    return $this->solr->createEndpoint($configuration, TRUE);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
   public function getFile($file = NULL) {
     $this->connect();
-
-    $query = $this->solr->createPing();
-    $query->setHandler('admin/file');
-    $query->addParam('contentType', 'text/xml;charset=utf-8');
+    $query = $this->solr->createApi([
+      'handler' => $this->configuration['core'] . '/admin/file',
+    ]);
     if ($file) {
       $query->addParam('file', $file);
     }
@@ -899,4 +1230,23 @@ abstract class SolrConnectorPluginBase extends ConfigurablePluginBase implements
     return parent::__sleep();
   }
 
+  /**
+   * {@inheritdoc}
+   */
+  public function alterConfigFiles(array &$files, string $lucene_match_version, string $server_id = '') {
+    if (!empty($this->configuration['jmx'])) {
+      $files['solrconfig_extra.xml'] .= "<jmx />\n";
+    }
+
+    if (!empty($this->configuration['jts'])) {
+      $jts_arguments = 'spatialContextFactory="org.locationtech.spatial4j.context.jts.JtsSpatialContextFactory" autoIndex="true" validationRule="repairBuffer0"';
+      $files['schema.xml'] = preg_replace("#\sclass\s*=\s*\"solr\.SpatialRecursivePrefixTreeFieldType\"#ms", "\\0\n        " . $jts_arguments, $files['schema.xml']);
+    }
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function alterConfigZip(ZipStream $zip, string $lucene_match_version, string $server_id = '') {
+  }
 }
