@@ -2,12 +2,18 @@
 
 namespace Drupal\Tests\search_api_solr\Functional;
 
+use Drupal\search_api\Entity\Index;
+use Drupal\search_api_solr\Utility\SolrCommitTrait;
+use Drupal\Tests\search_api\Functional\IntegrationTest as SearchApiIntegrationTest;
+
 /**
  * Tests the overall functionality of the Search API framework and admin UI.
  *
  * @group search_api_solr
  */
-class IntegrationTest extends \Drupal\Tests\search_api\Functional\IntegrationTest {
+class IntegrationTest extends SearchApiIntegrationTest {
+
+  use SolrCommitTrait;
 
   /**
    * The backend of the search server used for this test.
@@ -19,19 +25,19 @@ class IntegrationTest extends \Drupal\Tests\search_api\Functional\IntegrationTes
   /**
    * {@inheritdoc}
    */
-  public static $modules = array(
+  public static $modules = [
     'search_api_solr',
     'search_api_solr_test',
-  );
+  ];
 
   /**
    * {@inheritdoc}
    */
-  protected function tearDown() {
+  protected function tearDown(): void {
     if ($this->indexId) {
       if ($index = $this->getIndex()) {
         $index->clear();
-        sleep(2);
+        $this->ensureCommit($index);
       }
     }
     parent::tearDown();
@@ -102,8 +108,8 @@ class IntegrationTest extends \Drupal\Tests\search_api\Functional\IntegrationTes
     $edit += [
       'backend_config[connector_config][host]' => 'localhost',
       'backend_config[connector_config][port]' => '8983',
-      'backend_config[connector_config][path]' => '/',
-      'backend_config[connector_config][core]' => '',
+      'backend_config[connector_config][path]' => '/foo',
+      'backend_config[connector_config][core]' => 'bar',
     ];
     $this->submitForm($edit, 'Save');
 
@@ -117,8 +123,8 @@ class IntegrationTest extends \Drupal\Tests\search_api\Functional\IntegrationTes
     $edit = [
       'backend_config[connector_config][host]' => 'localhost',
       'backend_config[connector_config][port]' => '8983',
-      'backend_config[connector_config][path]' => '/solr',
-      'backend_config[connector_config][core]' => 'd8',
+      'backend_config[connector_config][path]' => '/',
+      'backend_config[connector_config][core]' => 'drupal',
     ];
     $this->submitForm($edit, 'Save');
     $this->assertSession()->pageTextContains('The Solr server could be reached.');
@@ -129,10 +135,13 @@ class IntegrationTest extends \Drupal\Tests\search_api\Functional\IntegrationTes
    *
    * @return int
    *   The number of successfully indexed items.
+   *
+   * @throws \Drupal\search_api\SearchApiException
    */
   protected function indexItems() {
     $index_status = parent::indexItems();
-    sleep(2);
+    $index = Index::load($this->indexId);
+    $this->ensureCommit($index);
     return $index_status;
   }
 
