@@ -2,6 +2,7 @@
 
 namespace Drupal\quickedit_test;
 
+use Drupal\Core\Access\AccessResult;
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\quickedit\Access\QuickEditEntityFieldAccessCheckInterface;
 
@@ -14,7 +15,19 @@ class MockQuickEditEntityFieldAccessCheck implements QuickEditEntityFieldAccessC
    * {@inheritdoc}
    */
   public function accessEditEntityField(EntityInterface $entity, $field_name) {
-    return TRUE;
+    switch (\Drupal::state()->get('quickedit_test_field_access')) {
+      case 'allowed':
+        return AccessResult::allowed();
+
+      case 'neutral':
+        return AccessResult::neutral();
+
+      case 'forbidden':
+        return AccessResult::forbidden();
+
+      default:
+        throw new \OutOfRangeException("The state for the 'quickedit_test_field_access' key must be either 'allowed', 'neutral' or 'forbidden'.");
+    }
   }
 
 }
