@@ -1,10 +1,53 @@
 <?php
 
 /**
- * Installs the standard highlighter config.
+ * @file
  */
-function search_api_solr_post_update_install_standard_highlighter_config() {
-  /** @var \Drupal\Core\Config\ConfigInstallerInterface $config_installer */
-  $config_installer = \Drupal::service('config.installer');
-  $config_installer->installDefaultConfig('module', 'search_api_solr');
+
+/**
+ * Delete Solr 4 and 5 field types.
+ */
+function search_api_solr_post_update_8204_replace_solr_4_field_types() {
+  try {
+    $storage = \Drupal::entityTypeManager()->getStorage('solr_field_type');
+    $storage->delete($storage->loadMultiple([
+      'm_text_und_5_2_0',
+      'text_und_4_5_0',
+      'm_text_de_5_2_0',
+      'm_text_en_5_2_0',
+      'm_text_nl_5_2_0',
+      'text_cs_5_0_0',
+      'text_de_4_5_0',
+      'text_de_5_0_0',
+      'text_de_scientific_5_0_0',
+      'text_el_4_5_0',
+      'text_en_4_5_0',
+      'text_es_4_5_0',
+      'text_fi_4_5_0',
+      'text_fr_4_5_0',
+      'text_it_4_5_0',
+      'text_nl_4_5_0',
+      'text_ru_4_5_0',
+      'text_uk_4_5_0',
+    ]));
+  }
+  catch (\Exception $e) {
+    // Don't break the upgrade, ignore the error because it is just nice to have
+    // cleanup.
+  }
+}
+
+/**
+ * Install new Solr Field Types and uninstall search_api_solr_multilingual.
+ */
+function search_api_solr_post_update_8319() {
+  if (\Drupal::moduleHandler()->moduleExists('search_api_solr_multilingual')) {
+    /** @var \Drupal\Core\Extension\ModuleInstallerInterface $module_installer */
+    $module_installer = \Drupal::service('module_installer');
+    $module_installer->uninstall(['search_api_solr_multilingual']);
+  }
+  // module_load_include is required in case that no update_hooks were run
+  // before.
+  module_load_include('install', 'search_api_solr');
+  search_api_solr_update_helper_install_configs();
 }
