@@ -4,7 +4,6 @@ namespace Drupal\search_api_solr_test\Logger;
 
 use Psr\Log\AbstractLogger;
 
-
 /**
  * A simple in memory logger.
  */
@@ -13,7 +12,7 @@ class InMemoryLogger extends AbstractLogger {
   private $messages = [];
 
   /**
-   *  {@inheritdoc}
+   * {@inheritdoc}
    */
   public function log($level, $message, array $context = []) {
     $this->messages[] = [
@@ -23,7 +22,11 @@ class InMemoryLogger extends AbstractLogger {
     ];
   }
 
+  /**
+   * Gets the last log message.
+   */
   public function getLastMessage() {
     return end($this->messages);
   }
+
 }
