@@ -3,27 +3,30 @@
 namespace Drupal\Tests\search_api_solr\Kernel\Processor;
 
 use Drupal\search_api\Entity\Server;
+use Drupal\search_api_solr\Utility\SolrCommitTrait;
+use Monolog\Handler\StreamHandler;
+use Monolog\Logger;
 use Symfony\Component\Yaml\Yaml;
+
+defined('SOLR_CLOUD') || define('SOLR_CLOUD', getenv('SOLR_CLOUD') ?: 'false');
 
 /**
  * Helper to exchange the DB backend for a Solr backend in processor tests.
  */
 trait SolrBackendTrait {
 
+  use SolrCommitTrait;
+
   /**
    * Swap the DB backend for a Solr backend.
    *
    * This function has to be called from the test setUp() function.
-   *
-   * @param string $module
-   *   The module that provides the server config.
-   * @param $config
-   *   The server config
    */
-  protected function enableSolrServer($module, $config) {
+  protected function enableSolrServer() {
+    $config = '/config/install/search_api.server.solr_search_server' . ('true' === SOLR_CLOUD ? '_cloud' : '') . '.yml';
     $this->server = Server::create(
       Yaml::parse(file_get_contents(
-        drupal_get_path('module', $module) . $config
+        drupal_get_path('module', 'search_api_solr_test') . $config
       ))
     );
     $this->server->save();
@@ -34,7 +37,7 @@ trait SolrBackendTrait {
     $index_storage = $this->container
       ->get('entity_type.manager')
       ->getStorage('search_api_index');
-    $index_storage->resetCache(array($this->index->id()));
+    $index_storage->resetCache([$this->index->id()]);
     $this->index = $index_storage->load($this->index->id());
   }
 
@@ -43,7 +46,7 @@ trait SolrBackendTrait {
    */
   protected function indexItems() {
     $index_status = parent::indexItems();
-    sleep(2);
+    $this->ensureCommit($this->index);
     return $index_status;
   }
 
@@ -52,7 +55,7 @@ trait SolrBackendTrait {
    */
   protected function tearDown() {
     $this->index->clear();
-    sleep(2);
+    $this->ensureCommit($this->index);
     parent::tearDown();
   }
 
